Docs: criando documentação para todos os Controllers

+++ src/main/java/com/ronyelison/quiz/service/ThemeService.java
@@ -7,6 +7,7 @@ import com.ronyelison.quiz.entity.Theme;
 import com.ronyelison.quiz.repository.ThemeRepository;
 import com.ronyelison.quiz.service.exception.ThemeNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,10 +30,18 @@ public class ThemeService {
     public void removeTheme(Long id){
         Theme theme = repository.findById(id)
                 .orElseThrow(() -> new ThemeNotFoundException("Tema não encontrado"));
+
+        if (theme.containsQuestionsInTheList()){
+            throw new DataIntegrityViolationException("Não é permitido remover um Tema que contém questões ligadas a ele");
+        }
         repository.delete(theme);
     }
 
     public List<ThemeResponse> findAllThemes(){
+        if (repository.findAll().isEmpty()){
+            throw new ThemeNotFoundException("Nenhum tema foi cadastrado");
+        }
+
         return repository.findAll()
                 .stream()
                 .map(Theme::entityToResponse)
